+++ handlers/model.py
@@ -16,8 +16,8 @@ from utils.db_requests import (
     )
 
 from utils.local_vars import (
-    available_work_buttons, 
     available_sites_buttons,
+    available_work_buttons,
     )
 
 from utils.functions import (
@@ -25,28 +25,13 @@ from utils.functions import (
     sum_for_week, 
     bonus, 
     for_next_bonus,
+    create_keyboard,
+    sites_cb,
+    get_keyboard,
     )
 from utils.settings import dp, bot
 
 current_date = date.today()
-
-sites_cb = CallbackData('vote', 'action')  # vote:<action>
-
-def get_keyboard():
-    keyboard_markup = types.InlineKeyboardMarkup(row_width=1)
-    row_btns_1 = (
-        types.InlineKeyboardButton(available_sites_buttons[0], callback_data=sites_cb.new(action=available_sites_buttons[0])),
-        types.InlineKeyboardButton(available_sites_buttons[1], callback_data=sites_cb.new(action=available_sites_buttons[1])),
-        types.InlineKeyboardButton(available_sites_buttons[2], callback_data=sites_cb.new(action=available_sites_buttons[2])),
-        )
-    row_btns_2 = (
-        types.InlineKeyboardButton(available_sites_buttons[3], callback_data=sites_cb.new(action=available_sites_buttons[3])),
-        types.InlineKeyboardButton(available_sites_buttons[4], callback_data=sites_cb.new(action=available_sites_buttons[4])),
-        types.InlineKeyboardButton(available_sites_buttons[5], callback_data=sites_cb.new(action=available_sites_buttons[5])),
-        )
-    keyboard_markup.row(*row_btns_1)
-    keyboard_markup.row(*row_btns_2)
-    return keyboard_markup
 
 async def model_deals(message):
     """
@@ -140,7 +125,10 @@ async def report_by_site(callback_query, state):
     await state.update_data(chosen_site=answer_data.lower())
     
     # await callback_query.answer(f'')
-    await bot.send_message(callback_query.from_user.id, f"Введите сумму для {answer_data}", reply_markup=types.ReplyKeyboardRemove())
+    await bot.send_message(
+        callback_query.from_user.id, f"Введите сумму для {answer_data}", 
+        reply_markup= types.ReplyKeyboardRemove()
+    )
 
     await OrderDeals.waiting_for_report_sum.set()
 
@@ -151,9 +139,7 @@ async def report_sum(message, state):
     Функция принимает сумму для записи в бд
     """
 
-    markup = types.ReplyKeyboardMarkup(resize_keyboard=True)
-    for button in available_work_buttons:
-        markup.add(button)
+    markup = create_keyboard(available_work_buttons)
 
 
     report_sum = message.text
@@ -163,6 +149,7 @@ async def report_sum(message, state):
         site = await state.get_data()
         site = site['chosen_site']
 
+        # Запись суммы в БД
         with sqlite3.connect("database.db") as conn:
             name = name_by_chat_id(conn, chat_id)
             name = name[0]
@@ -180,5 +167,5 @@ async def report_sum(message, state):
     
     except Exception as e:
         logger.error(f"Ошибка ввода суммы: {e}")
-        await message.answer("Сумма должна быть числом")
+        await message.answer("Сумма должна быть числом", reply_markup=markup)
         await OrderDeals.waiting_for_modeldeals.set()
